refactor: expo plugins
