+++ plugin/src/plugin-props.type.ts
@@ -1,0 +1,4 @@
+export interface PluginProps {
+  bonjourServicesName?: string;
+  localNetworkUsagePermissionText?: string;
+}
